refactor: 테스트 클래스 패키지 경로 정리 및 불필요한 import 제거
