+++ backend/src/test/java/coffeeshout/cardgame/domain/CardHandTest.java
@@ -1,11 +1,8 @@
-package coffeeshout.cardgame.domain.cardgame;
+package coffeeshout.cardgame.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import coffeeshout.cardgame.domain.CardGameRound;
-import coffeeshout.cardgame.domain.CardGameScore;
-import coffeeshout.cardgame.domain.CardHand;
 import coffeeshout.cardgame.domain.card.AdditionCard;
 import coffeeshout.cardgame.domain.card.Card;
 import coffeeshout.cardgame.domain.card.MultiplierCard;
