+++ backend/src/test/java/coffeeshout/cardgame/domain/card/DeckTest.java
@@ -1,12 +1,8 @@
-package coffeeshout.cardgame.domain.cardgame.card;
+package coffeeshout.cardgame.domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import coffeeshout.cardgame.domain.card.AdditionCard;
-import coffeeshout.cardgame.domain.card.Card;
-import coffeeshout.cardgame.domain.card.Deck;
-import coffeeshout.cardgame.domain.card.MultiplierCard;
 import java.util.List;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.BeforeEach;
