+++ backend/src/test/java/coffeeshout/racinggame/ui/RacingGameIntegrationTest.java
@@ -1,4 +1,4 @@
-package coffeeshout.minigame.racinggame.ui;
+package coffeeshout.racinggame.ui;
 
 import coffeeshout.fixture.RoomFixture;
 import coffeeshout.fixture.TestStompSession;
